[begin] 코드 분리
