+++ src/IceServersHandler.ts
@@ -1,0 +1,26 @@
+// IceServersHandler.js
+
+const IceServersHandler = (function () {
+  function getIceServers(connection) {
+    // resiprocate: 3344+4433
+    // pions: 7575
+    var iceServers = [
+      {
+        urls: [
+          "stun:stun.l.google.com:19302",
+          "stun:stun1.l.google.com:19302",
+          "stun:stun2.l.google.com:19302",
+          "stun:stun.l.google.com:19302?transport=udp",
+        ],
+      },
+    ];
+
+    return iceServers;
+  }
+
+  return {
+    getIceServers: getIceServers,
+  };
+})();
+
+export default IceServersHandler;
