+++ tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "ESNext",
     "useDefineForClassFields": true,
+    "moduleResolution": "Node",
     "module": "ESNext",
     "lib": ["ESNext", "DOM"],
     "strict": true,
